Refactored to MVC
